social_media_api/requirements.txt
asgiref
Django
django-taggit
djangorestframework
djangorestframework-simplejwt
psycopg2
PyJWT
sqlparse
tzdata
django-filter
psycopg2-binary
django-sslserver
waitress
django-cors-headers
Pillow